feat: ディレクトリ構造ガイドを更新し、テストファイルのパスを修正、日付フィルタリング関数のテストを追加
