+++ eslint.config.mjs
@@ -78,7 +78,7 @@ const eslintConfig = tseslint.config(
   },
   {
     extends: [unicornPlugin.configs.recommended],
-    ignores: ['tests/**/*.test.{ts,tsx}', 'tests/**/*.spec.{ts,tsx}'],
+    ignores: ['src/tests/**/*.test.{ts,tsx}', 'src/tests/**/*.spec.{ts,tsx}'],
     rules: {
       'unicorn/prevent-abbreviations': 'off',
     },
